move all logic to internal package
