+++ internal/cmd/response_writers.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"bytes"
